+++ commcare_connect/organization/models.py
@@ -1,0 +1,47 @@
+from django.conf import settings
+from django.db import models
+from django.utils.translation import gettext_lazy as _
+
+from commcare_connect.users.models import User
+from commcare_connect.utils.db import BaseModel, slugify_uniquely
+
+
+class Organization(BaseModel):
+    name = models.CharField(max_length=255)
+    slug = models.SlugField(max_length=255, unique=True)
+    members = models.ManyToManyField(
+        settings.AUTH_USER_MODEL, related_name="organizations", through="UserOrganizationMembership"
+    )
+
+    def save(self, *args, **kwargs):
+        if not self.id:
+            self.slug = slugify_uniquely(self.name, self.__class__)
+        super().save(*args, **kwargs)
+
+    def __str__(self):
+        return self.slug
+
+    class Meta:
+        managed = False
+
+
+class UserOrganizationMembership(models.Model):
+    class Role(models.TextChoices):
+        ADMIN = "admin", _("Admin")
+        MEMBER = "member", _("Member")
+
+    organization = models.ForeignKey(
+        Organization,
+        on_delete=models.CASCADE,
+        related_name="memberships",
+    )
+    user = models.ForeignKey(
+        User,
+        on_delete=models.DO_NOTHING,
+        related_name="memberships",
+    )
+    role = models.CharField(max_length=20, choices=Role.choices, default=Role.MEMBER)
+
+    class Meta:
+        db_table = "organization_membership"
+        managed = False
