+++ commcare_connect/users/admin.py
@@ -3,8 +3,8 @@ from django.contrib.auth import admin as auth_admin
 from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
 
+from commcare_connect.organization.models import Organization, UserOrganizationMembership
 from commcare_connect.users.forms import OrganizationCreationForm, UserAdminChangeForm, UserAdminCreationForm
-from commcare_connect.users.models import Organization, UserOrganizationMembership
 
 User = get_user_model()
